+++ Nagru - Manga Organizer/Properties/Settings.Designer.cs
@@ -77,7 +77,7 @@ namespace Nagru___Manga_Organizer.Properties {
         
         [global::System.Configuration.ApplicationScopedSettingAttribute()]
         [global::System.Diagnostics.DebuggerNonUserCodeAttribute()]
-        [global::System.Configuration.DefaultSettingValueAttribute("3.8.00")]
+        [global::System.Configuration.DefaultSettingValueAttribute("3.8.12")]
         public string Version {
             get {
                 return ((string)(this["Version"]));